Time: 0 ms (100.00%) | Memory: 7.8 MB (65.39%) - LeetSync

+++ 202-happy-number/happy-number.cpp
@@ -9,7 +9,7 @@ public:
             return false;
         }
         }
-        long long int sum=0;
+        int sum=0;
         while(n>0){
             int dig=n%10;
             sum+=dig*dig;
